+++ tests/testthat/test-check_collinearity.R
@@ -1,3 +1,28 @@
+# https://github.com/easystats/performance/pull/547
+test_that("check_collinearity, correct order in print", {
+  m <- lm(mpg ~ wt + cyl + gear + disp, data = mtcars)
+  out <- capture.output(print(check_collinearity(m)))
+  expect_identical(
+    out,
+    c(
+      "# Check for Multicollinearity",
+      "",
+      "Low Correlation",
+      "",
+      " Term  VIF    VIF 95% CI Increased SE Tolerance Tolerance 95% CI",
+      " gear 1.53 [1.19,  2.51]         1.24      0.65     [0.40, 0.84]",
+      "",
+      "Moderate Correlation",
+      "",
+      " Term  VIF    VIF 95% CI Increased SE Tolerance Tolerance 95% CI",
+      "   wt 5.05 [3.21,  8.41]         2.25      0.20     [0.12, 0.31]",
+      "  cyl 5.41 [3.42,  9.04]         2.33      0.18     [0.11, 0.29]",
+      " disp 9.97 [6.08, 16.85]         3.16      0.10     [0.06, 0.16]"
+    )
+  )
+})
+
+
 if (requiet("glmmTMB") && getRversion() >= "4.0.0") {
   data(Salamanders)
   m1 <- glmmTMB(count ~ spp + mined + (1 | site),
@@ -59,7 +84,7 @@ if (requiet("glmmTMB") && getRversion() >= "4.0.0") {
       tolerance = 1e-3
     )
 
-    suppressWarnings(coll <- check_collinearity(m2, component = "all", verbose = FALSE))
+    suppressWarnings(coll <- check_collinearity(m2, component = "all", verbose = FALSE)) # nolint
     expect_true(all(coll$Tolerance < coll$Tolerance_CI_high))
     expect_true(all(coll$VIF > coll$VIF_CI_low))
 
@@ -92,9 +117,9 @@ if (requiet("glmmTMB") && getRversion() >= "4.0.0") {
         )
       }))
 
-      expect_message(ccoM <- check_collinearity(aM))
-      expect_warning(expect_message(ccoW <- check_collinearity(aW)))
-      expect_message(ccoB <- check_collinearity(aB), regexp = NA)
+      expect_message(ccoM <- check_collinearity(aM)) # nolint
+      expect_warning(expect_message(ccoW <- check_collinearity(aW))) # nolint
+      expect_message(ccoB <- check_collinearity(aB), regexp = NA) # nolint
 
       expect_identical(nrow(ccoM), 15L)
       expect_identical(nrow(ccoW), 3L)
@@ -117,9 +142,9 @@ if (requiet("glmmTMB") && getRversion() >= "4.0.0") {
         )
       }))
 
-      expect_message(ccoM <- check_collinearity(aM))
-      expect_warning(expect_message(ccoW <- check_collinearity(aW)))
-      expect_message(ccoB <- check_collinearity(aB), regexp = NA)
+      expect_message(ccoM <- check_collinearity(aM)) # nolint
+      expect_warning(expect_message(ccoW <- check_collinearity(aW))) # nolint
+      expect_message(ccoB <- check_collinearity(aB), regexp = NA) # nolint
 
       expect_identical(nrow(ccoM), 15L)
       expect_identical(nrow(ccoW), 3L)